reorganized App, cleaned up routes, created final page/component relationships
