+++ client/src/components/Sidebar/Sidebar.jsx
@@ -1,9 +1,11 @@
-const { ref } = require('joi');
-
 import React from 'react';
 
 const Sidebar = () => {
-  return <div></div>;
+  return (
+    <div>
+      <h1>This is the Sidebar</h1>
+    </div>
+  );
 };
 
 export default Sidebar;
