+++ client/src/components/Widgets/Widgets.jsx
@@ -1,7 +1,11 @@
 import React from 'react';
 
 const Widgets = () => {
-  return <div></div>;
+  return (
+    <div>
+      <h1>WIdgets will go here</h1>
+    </div>
+  );
 };
 
 export default Widgets;
