+++ client/src/pages/AppBody/AppBody.jsx
@@ -1,11 +1,18 @@
-import React from 'react'
+import React from 'react';
+import Header from '../../components/Header/Header';
+import Newsfeed from '../../components/Newsfeed/Newsfeed';
+import Sidebar from '../../components/Sidebar/Sidebar';
+import Widgets from '../../components/Widgets/Widgets';
 
-const AppBody = () {
-    return (
-        <div>
-            
-        </div>
-    )
-}; 
+const AppBody = () => {
+  return (
+    <div>
+      <Header />
+      <Newsfeed />
+      <Sidebar />
+      <Widgets />
+    </div>
+  );
+};
 
-export default AppBody; 
+export default AppBody;
