+++ client/src/pages/LoginSignup/LoginSignup.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const LoginSignup = () => {
+  return <h1>This is the Login/Signup Page</h1>;
+};
+
+export default LoginSignup;
